Move coordinator to separate module in smart_meter_texas
