+++ homeassistant/components/smart_meter_texas/__init__.py
@@ -1,30 +1,18 @@
 """The Smart Meter Texas integration."""
 
 import logging
-import ssl
 
-from smart_meter_texas import Account, Client
-from smart_meter_texas.exceptions import (
-    SmartMeterTexasAPIError,
-    SmartMeterTexasAuthError,
-)
+from smart_meter_texas import Account
+from smart_meter_texas.exceptions import SmartMeterTexasAuthError
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_PASSWORD, CONF_USERNAME, Platform
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryNotReady
-from homeassistant.helpers import aiohttp_client
-from homeassistant.helpers.debounce import Debouncer
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 from homeassistant.util.ssl import get_default_context
 
-from .const import (
-    DATA_COORDINATOR,
-    DATA_SMART_METER,
-    DEBOUNCE_COOLDOWN,
-    DOMAIN,
-    SCAN_INTERVAL,
-)
+from .const import DATA_COORDINATOR, DATA_SMART_METER, DOMAIN
+from .coordinator import SmartMeterTexasCoordinator, SmartMeterTexasData
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -52,26 +40,7 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
 
     await smart_meter_texas_data.setup()
 
-    async def async_update_data():
-        _LOGGER.debug("Fetching latest data")
-        await smart_meter_texas_data.read_meters()
-        return smart_meter_texas_data
-
-    # Use a DataUpdateCoordinator to manage the updates. This is due to the
-    # Smart Meter Texas API which takes around 30 seconds to read a meter.
-    # This avoids Home Assistant from complaining about the component taking
-    # too long to update.
-    coordinator = DataUpdateCoordinator(
-        hass,
-        _LOGGER,
-        config_entry=entry,
-        name="Smart Meter Texas",
-        update_method=async_update_data,
-        update_interval=SCAN_INTERVAL,
-        request_refresh_debouncer=Debouncer(
-            hass, _LOGGER, cooldown=DEBOUNCE_COOLDOWN, immediate=True
-        ),
-    )
+    coordinator = SmartMeterTexasCoordinator(hass, entry, smart_meter_texas_data)
 
     hass.data.setdefault(DOMAIN, {})
     hass.data[DOMAIN][entry.entry_id] = {
@@ -86,38 +55,6 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     await hass.config_entries.async_forward_entry_setups(entry, PLATFORMS)
 
     return True
-
-
-class SmartMeterTexasData:
-    """Manages coordinatation of API data updates."""
-
-    def __init__(
-        self,
-        hass: HomeAssistant,
-        entry: ConfigEntry,
-        account: Account,
-        ssl_context: ssl.SSLContext,
-    ) -> None:
-        """Initialize the data coordintator."""
-        self._entry = entry
-        self.account = account
-        websession = aiohttp_client.async_get_clientsession(hass)
-        self.client = Client(websession, account, ssl_context=ssl_context)
-        self.meters: list = []
-
-    async def setup(self):
-        """Fetch all of the user's meters."""
-        self.meters = await self.account.fetch_meters(self.client)
-        _LOGGER.debug("Discovered %s meter(s)", len(self.meters))
-
-    async def read_meters(self):
-        """Read each meter."""
-        for meter in self.meters:
-            try:
-                await meter.read_meter(self.client)
-            except (SmartMeterTexasAPIError, SmartMeterTexasAuthError) as error:
-                raise UpdateFailed(error) from error
-        return self.meters
 
 
 async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
